Player Movement and Boat Movement
I made the player be able to move sprint and jump. The player is a first person perspective. the boat has a birds eye view when the player is on the wheel. The boat turns left and right when the boat is moving.
I recommend changing the cameras to Cinemachine but Unity 6 completely changed it so youtube videos are outdated.

The Only thing I couldn't figure out is how to make the player rotate along with the boat while it is turning. I tried for hours but I just couldn't get it