Added another Event for better underwater logic
Added another event to the UnderWaterTrigger script, which checks if the players head is submerged. (Added a new box collider to the player)

Overhauled water logic in the PlayerManager script so the swimming and floating bools are accurately enabled and disabled. And the other logic in there.

UnderwaterTransition now activates based on the head submerge/emerge events instead of the body events.