I added the start and the ending for the game.
I used player prefs to make all of the settings save. I also added models for the Fusion Forge and ATMOS Terminal so it looks like they were made out of old appliances. I added the human to talk to, who tells us the objective of the game. When you return to her after delivering all of the packages, it shows the ending of the game and can return to menu or credits.