Fixing water animals
I believe I fixed water animals so now they choose a spot too move to it has to be in the ocean and it when it leaves the water it immediately stops and searches for a new spot. It also says the spot is invalid if it hits the standable layer. So i believe they can no longer go on land